+++ docs/changes.rst
@@ -18,6 +18,7 @@ Released on ? 2018.
 * Integrated maintenance and cleanup tasks using Celery.
 * Improved performance of loading translations by almost 25%.
 * Removed support for merging headers on upload.
+* Removed support for custom commit messages.
 
 weblate 3.1.1
 -------------